requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
unstructured
python-dotenv
google-generativeai
PyPDF2
faiss-cpu
